Add Simple chunking mechanism. subdivision terrain generation
Decided to use full height chunks, that will be loaded and unload on demand. Also added simple algo to create terrain by subdivision. Gonna expand on that concept tomorrow. I expect to use the subdivison mechanism to generate hills+ valleys, use noise to generate POI ( noise is deterministic, it will always result in same number for particular location, only way to modify it is the seed.